0.7.0
feat/refactor/style - Modification de la base de donnees, ajout d'une nouvelle relation et d'une table category_product, Ajout choices.js, ajout vue des produits, ajout modification des produits, assignation à plusieurs catégories, corrections UI

+++ database/migrations/2023_03_02_090359_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('label', 150);
-            $table->integer('price_unit');
+            $table->decimal('price_unit', 8, 2);
             $table->longText('picture_path');
             $table->string('color', 20);
             $table->string('brand', 20);
